fixed event_object_movement constants by expanding movementActionId by 8 bytes. 'THIS WILL BREAK EXISTING SAVES'

+++ include/field_player_avatar.h
@@ -43,7 +43,7 @@ void SetPlayerAvatarFieldMove(void);
 u16 GetPlayerAvatarGraphicsIdByCurrentState(void);
 void SetPlayerAvatarStateMask(u8 flags);
 u16 GetPlayerAvatarGraphicsIdByStateId(u8 state);
-u8 GetJumpSpecialMovementAction(u32);
+u16 GetJumpSpecialMovementAction(u32);
 bool8 PartyHasMonWithSurf(void);
 bool8 IsPlayerFacingSurfableFishableWater(void);
 bool8 IsPlayerSurfingNorth(void);
